refactor(wip): make plugins more low-level
