+++ addon/index.ts
@@ -1,5 +1,9 @@
 export { default as RdfaEditorPlugin } from '@lblod/ember-rdfa-editor/core/rdfa-editor-plugin';
 export { NodeSpec } from 'prosemirror-model';
-export { getRdfaAttrs, rdfaAttrs } from '@lblod/ember-rdfa-editor/core/schema';
+export {
+  rdfaSchema,
+  getRdfaAttrs,
+  rdfaAttrs,
+} from '@lblod/ember-rdfa-editor/core/schema';
 export { Plugin as ProsePlugin, PluginKey } from 'prosemirror-state';
 export { DecorationSet, Decoration } from 'prosemirror-view';
