+++ addon/marks/em.ts
@@ -1,0 +1,8 @@
+import { MarkSpec } from 'prosemirror-model';
+
+export const em: MarkSpec = {
+  parseDOM: [{ tag: 'i' }, { tag: 'em' }, { style: 'font-style=italic' }],
+  toDOM() {
+    return ['em', 0];
+  },
+};
