+++ addon/marks/strikethrough.ts
@@ -1,0 +1,8 @@
+import { MarkSpec } from 'prosemirror-model';
+
+export const strikethrough: MarkSpec = {
+  parseDOM: [{ tag: 's' }, { tag: 'del' }],
+  toDOM() {
+    return ['del', 0];
+  },
+};
