+++ addon/marks/underline.ts
@@ -1,0 +1,8 @@
+import { MarkSpec } from 'prosemirror-model';
+
+export const underline: MarkSpec = {
+  parseDOM: [{ tag: 'u' }],
+  toDOM() {
+    return ['u', 0];
+  },
+};
