+++ addon/nodes/blockquote.ts
@@ -1,0 +1,11 @@
+import { NodeSpec } from 'prosemirror-model';
+
+export const blockquote: NodeSpec = {
+  content: 'block+',
+  group: 'block',
+  defining: true,
+  parseDOM: [{ tag: 'blockquote' }],
+  toDOM() {
+    return ['blockquote', 0];
+  },
+};
