+++ addon/nodes/code-block.ts
@@ -1,0 +1,13 @@
+import { NodeSpec } from 'prosemirror-model';
+
+export const code_block: NodeSpec = {
+  content: 'text*',
+  marks: '',
+  group: 'block',
+  code: true,
+  defining: true,
+  parseDOM: [{ tag: 'pre', preserveWhitespace: 'full' }],
+  toDOM() {
+    return ['pre', ['code', 0]];
+  },
+};
