+++ addon/nodes/doc.ts
@@ -1,0 +1,5 @@
+import { NodeSpec } from 'prosemirror-model';
+
+export const doc: NodeSpec = {
+  content: 'block+',
+};
