+++ addon/nodes/hard-break.ts
@@ -1,0 +1,11 @@
+import { NodeSpec } from 'prosemirror-model';
+
+export const hard_break: NodeSpec = {
+  inline: true,
+  group: 'inline',
+  selectable: false,
+  parseDOM: [{ tag: 'br' }],
+  toDOM() {
+    return ['br'];
+  },
+};
