+++ addon/nodes/horizontal-rule.ts
@@ -1,0 +1,9 @@
+import { NodeSpec } from 'prosemirror-model';
+
+export const horizontal_rule: NodeSpec = {
+  group: 'block',
+  parseDOM: [{ tag: 'hr' }],
+  toDOM() {
+    return ['hr'];
+  },
+};
