+++ addon/nodes/image.ts
@@ -1,0 +1,27 @@
+import { Node as PNode, NodeSpec } from 'prosemirror-model';
+
+export const image: NodeSpec = {
+  inline: true,
+  attrs: {
+    src: {},
+    alt: { default: null },
+    title: { default: null },
+  },
+  group: 'inline',
+  draggable: true,
+  parseDOM: [
+    {
+      tag: 'img[src]',
+      getAttrs(dom: HTMLElement) {
+        return {
+          src: dom.getAttribute('src'),
+          title: dom.getAttribute('title'),
+          alt: dom.getAttribute('alt'),
+        };
+      },
+    },
+  ],
+  toDOM(node: PNode) {
+    return ['img', node.attrs];
+  },
+};
