+++ addon/nodes/text.ts
@@ -1,0 +1,5 @@
+import { NodeSpec } from 'prosemirror-model';
+
+export const text: NodeSpec = {
+  group: 'inline',
+};
