+++ tests/dummy/app/controllers/index.ts
@@ -1,21 +1,26 @@
 import Controller from '@ember/controller';
 import { action } from '@ember/object';
-import { tracked, TrackedSet } from 'tracked-built-ins';
+import { tracked } from 'tracked-built-ins';
 import { ProseController } from '@lblod/ember-rdfa-editor/core/prosemirror';
+import { rdfaSchema } from '@lblod/ember-rdfa-editor';
+import { Plugin } from 'prosemirror-state';
+import { Schema } from 'prosemirror-model';
 
 export default class IndexController extends Controller {
   @tracked rdfaEditor?: ProseController;
-  @tracked plugins = new TrackedSet([
-    'code-mark',
-    'inline-components',
-    'table',
-    {
-      name: 'highlight',
-      options: {
-        testKey: 'test',
-      },
-    },
-  ]);
+  // @tracked plugins = new TrackedSet([
+  //   'code-mark',
+  //   'inline-components',
+  //   // 'table',
+  //   {
+  //     name: 'highlight',
+  //     options: {
+  //       testKey: 'test',
+  //     },
+  //   },
+  // ]);
+  @tracked plugins: Plugin[] = [];
+  schema: Schema = rdfaSchema;
 
   @action
   rdfaEditorInit(rdfaEditor: ProseController) {
