+++ tests/dummy/app/templates/index.hbs
@@ -16,6 +16,7 @@
     showIndentButtons='true'
   }}
   @plugins={{this.plugins}}
+  @schema={{this.schema}}
 >
   App version:
   {{app-version versionOnly=true showExtended=true}}
